universeterminalv1/backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
slowapi